% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alabamablock20.R
\docType{data}
\name{alabamablock20}
\alias{alabamablock20}
\title{Alabama}
\format{
A data frame with 3925 rows and 398 variables:
}
\source{
/url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(alabamablock20)
}
\description{
Blocks for the state of Alabama
}
\keyword{datasets}
