% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arkansasblock20.R
\docType{data}
\name{arkansasblock20}
\alias{arkansasblock20}
\title{Arkansas}
\format{
A data frame with 2294 rows and 398 variables:
}
\source{
/url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(arkansastract20)
}
\description{
Blocks for the state of Arkansas
}
\keyword{datasets}
