% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/californiablock20.R
\docType{data}
\name{californiablock20}
\alias{californiablock20}
\title{California}
\format{
A data frame with 25607 rows and 398 variables:
}
\source{
/url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(californiatract20)
}
\description{
Blocks for the state of California
}
\keyword{datasets}
