% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloradoblock20.R
\docType{data}
\name{coloradoblock20}
\alias{coloradoblock20}
\title{Colorado}
\format{
A data frame with  4058 rows and 398 variables:
}
\source{
/url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(alabamatract20)
}
\description{
Blocks for the state of Colorado
}
\keyword{datasets}
