% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connecticutblock20.R
\docType{data}
\name{connecticutblock20}
\alias{connecticutblock20}
\title{Connecticut}
\format{
A data frame with 2716 rows and 398 variables:
}
\source{
/url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(connecticutblock20)
}
\description{
Blocks for the state of Connecticut
}
\keyword{datasets}
