% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcblock20.R
\docType{data}
\name{dcblock20}
\alias{dcblock20}
\title{District of columbia}
\format{
A data frame with 571 rows and 398 variables:
}
\source{
/url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(dcblock20)
}
\description{
Blocks for District of columbia
}
\keyword{datasets}
