% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delawareblock20.R
\docType{data}
\name{delawareblock20}
\alias{delawareblock20}
\title{Delaware}
\format{
A data frame with 706 rows and 398 variables:
}
\source{
/url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(delawareblock20)
}
\description{
Blocks for the state of Delaware
}
\keyword{datasets}
