% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floridablock20.R
\docType{data}
\name{floridablock20}
\alias{floridablock20}
\title{Florida}
\format{
A data frame with 13388 rows and 398 variables:
}
\source{
/url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(floridablock20)
}
\description{
Blocks for the state of Florida
}
\keyword{datasets}
