% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/georgiablock20.R
\docType{data}
\name{georgiablock20}
\alias{georgiablock20}
\title{Georgia}
\format{
A data frame with 7446 rows and 398 variables:
}
\source{
/url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(georgiablock20)
}
\description{
Blocks for the state of Georgia
}
\keyword{datasets}
