% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hawaiiblock20.R
\docType{data}
\name{hawaiiblock20}
\alias{hawaiiblock20}
\title{Hawaii}
\format{
A data frame with 1083 rows and 398 variables:
}
\source{
/url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(hawaiiblock20)
}
\description{
Blocks for the state of Hawaii
}
\keyword{datasets}
