% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idahoblock20.R
\docType{data}
\name{idahoblock20}
\alias{idahoblock20}
\title{Idaho}
\format{
A data frame with 1284 rows and 398 variables:
}
\source{
/url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(idahoblock20)
}
\description{
Blocks for the state of Idaho
}
\keyword{datasets}
