% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/illinoisblock20.R
\docType{data}
\name{illinoisblock20}
\alias{illinoisblock20}
\title{Illinois}
\format{
A data frame with 9898 rows and 398 variables:
}
\source{
/url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(illinoisblock20)
}
\description{
Blocks for the state of Illinois
}
\keyword{datasets}
