% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kansasblock20.R
\docType{data}
\name{kansasblock20}
\alias{kansasblock20}
\title{Kansas}
\format{
A data frame with 2461 rows and 398 variables:
}
\source{
/url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(kansasblock20)
}
\description{
Blocks for the state of Kansas
}
\keyword{datasets}
