% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kentuckyblock20.R
\docType{data}
\name{kentuckyblock20}
\alias{kentuckyblock20}
\title{Kentucky}
\format{
A data frame with 3581 rows and 398 variables:
}
\source{
/url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(kentuckyblock20)
}
\description{
Blocks for the state of Kentucky
}
\keyword{datasets}
