% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/louisianablock20.R
\docType{data}
\name{louisianablock20}
\alias{louisianablock20}
\title{Louisiana}
\format{
A data frame with 4294 rows and 398 variables:
}
\source{
/url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(louisianablock20)
}
\description{
Blocks for the state of Louisiana
}
\keyword{datasets}
