% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maineblock20.R
\docType{data}
\name{maineblock20}
\alias{maineblock20}
\title{Maine}
\format{
A data frame with 1184 rows and 398 variables:
}
\source{
/url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(maineblock20)
}
\description{
Blocks for the state of Maine
}
\keyword{datasets}
