% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marylandblock20.R
\docType{data}
\name{marylandblock20}
\alias{marylandblock20}
\title{Maryland}
\format{
A data frame with 4079 rows and 398 variables:
}
\source{
/url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(marylandblock20)
}
\description{
Blocks for the state of Maryland
}
\keyword{datasets}
