% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/michiganblock20.R
\docType{data}
\name{michiganblock20}
\alias{michiganblock20}
\title{Michigan}
\format{
A data frame with 8386 rows and 398 variables:
}
\source{
/url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(michiganblock20)
}
\description{
Blocks for the state of Michigan
}
\keyword{datasets}
