% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minnesotablock20.R
\docType{data}
\name{minnesotablock20}
\alias{minnesotablock20}
\title{Minnesota}
\format{
A data frame with 4706 rows and 398 variables:
}
\source{
/url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(minnesotablock20)
}
\description{
Blocks for the state of Minnesota
}
\keyword{datasets}
