% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mississipiblock20.R
\docType{data}
\name{mississipiblock20}
\alias{mississipiblock20}
\title{Mississipi}
\format{
A data frame with 2445 rows and 398 variables:
}
\source{
/url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(mississipiblock20)
}
\description{
Blocks for the state of Mississipi
}
\keyword{datasets}
