% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nebraskablock20.R
\docType{data}
\name{nebraskablock20}
\alias{nebraskablock20}
\title{Nebraska}
\format{
A data frame with 1648 rows and 398 variables:
}
\source{
/url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(nebraskablock20)
}
\description{
Blocks for the state of Nebraska
}
\keyword{datasets}
