% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nevadablock20.R
\docType{data}
\name{nevadablock20}
\alias{nevadablock20}
\title{Nevada}
\format{
A data frame with 1963 rows and 398 variables:
}
\source{
/url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(nevadablock20)
}
\description{
Blocks for the state of Nevada
}
\keyword{datasets}
