% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newjerseyblock20.R
\docType{data}
\name{newjerseyblock20}
\alias{newjerseyblock20}
\title{New Jersey}
\format{
A data frame with 6599 rows and 398 variables:
}
\source{
/url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(newjerseyblock20)
}
\description{
Blocks for the state of New Jersey
}
\keyword{datasets}
