% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newmexicoblock20.R
\docType{data}
\name{newmexicoblock20}
\alias{newmexicoblock20}
\title{New Mexico}
\format{
A data frame with 1614 rows and 398 variables:
}
\source{
/url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(newmexicoblock20)
}
\description{
Blocks for the state of New Mexico
}
\keyword{datasets}
