% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newyorkblock20.R
\docType{data}
\name{newyorkblock20}
\alias{newyorkblock20}
\title{New York}
\format{
A data frame with 16070 rows and 398 variables:
}
\source{
/url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(newyorkblock20)
}
\description{
Blocks for the state of New York
}
\keyword{datasets}
