% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/northdakotablock20.R
\docType{data}
\name{northdakotablock20}
\alias{northdakotablock20}
\title{North Dakota}
\format{
A data frame with 632 rows and 398 variables:
}
\source{
/url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(northdakotablock20)
}
\description{
Blocks for the state of North Dakota
}
\keyword{datasets}
