% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohioblock20.R
\docType{data}
\name{ohioblock20}
\alias{ohioblock20}
\title{Ohio}
\format{
A data frame with 9472 rows and 398 variables:
}
\source{
/url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(ohioblock20)
}
\description{
Blocks for the state of Ohio
}
\keyword{datasets}
