% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oklahomablock20.R
\docType{data}
\name{oklahomablock20}
\alias{oklahomablock20}
\title{Oklahoma}
\format{
A data frame with 3374 rows and 398 variables:
}
\source{
/url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(oklahomablock20)
}
\description{
Blocks for the state of Oklahoma
}
\keyword{datasets}
