% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oregonblock20.R
\docType{data}
\name{oregonblock20}
\alias{oregonblock20}
\title{Oregon}
\format{
A data frame with 2970 rows and 398 variables:
}
\source{
/url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(oregonblock20)
}
\description{
Blocks for the state of Oregon
}
\keyword{datasets}
