% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pennsylvaniablock20.R
\docType{data}
\name{pennsylvaniablock20}
\alias{pennsylvaniablock20}
\title{Pennsylvania}
\format{
A data frame with 10173 rows and 398 variables:
}
\source{
/url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(pennsylvaniablock20)
}
\description{
Blocks for the state of Pennsylvania
}
\keyword{datasets}
