% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puertoricoblock20.R
\docType{data}
\name{puertoricoblock20}
\alias{puertoricoblock20}
\title{Puerto Rico}
\format{
A data frame with 2555 rows and 398 variables:
}
\source{
/url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(puertoricoblock20)
}
\description{
Blocks for the state of Puerto Rico
}
\keyword{datasets}
