% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhodeislandblock20.R
\docType{data}
\name{rhodeislandblock20}
\alias{rhodeislandblock20}
\title{Rhode Island}
\format{
A data frame with 792 rows and 398 variables:
}
\source{
/url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(rhodeislandblock20)
}
\description{
Blocks for the state of Rhode Island
}
\keyword{datasets}
