% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/southcarolinablock20.R
\docType{data}
\name{southcarolinablock20}
\alias{southcarolinablock20}
\title{South Carolina}
\format{
A data frame with 3408 rows and 398 variables:
}
\source{
/url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(southcarolinablock20)
}
\description{
Blocks for the state of South Carolina
}
\keyword{datasets}
