% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tennesseblock20.R
\docType{data}
\name{tennesseblock20}
\alias{tennesseblock20}
\title{Tennesse}
\format{
A data frame with 4562 rows and 398 variables:
}
\source{
/url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(tennesseblock20)
}
\description{
Blocks for the state of Tennesse
}
\keyword{datasets}
