% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texasblock20.R
\docType{data}
\name{texasblock20}
\alias{texasblock20}
\title{Texas}
\format{
A data frame with 18638 rows and 398 variables:
}
\source{
/url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(texasblock20)
}
\description{
Blocks for the state of Texas
}
\keyword{datasets}
