% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vermontblock20.R
\docType{data}
\name{vermontblock20}
\alias{vermontblock20}
\title{Vermont}
\format{
A data frame with 552 rows and 398 variables:
}
\source{
/url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(vermontblock20)
}
\description{
Blocks for the state of Vermont
}
\keyword{datasets}
