% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/washingtonblock20.R
\docType{data}
\name{washingtonblock20}
\alias{washingtonblock20}
\title{Washington}
\format{
A data frame with 5311 rows and 398 variables:
}
\source{
/url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(washingtonblock20)
}
\description{
Blocks for the state of Washington
}
\keyword{datasets}
