% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/westvirginiablock20.R
\docType{data}
\name{westvirginiablock20}
\alias{westvirginiablock20}
\title{West Virginia}
\format{
A data frame with 1639 rows and 398 variables:
}
\source{
/url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(westvirginiablock20)
}
\description{
Blocks for the state of West Virginia
}
\keyword{datasets}
