% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wisconsinblock20.R
\docType{data}
\name{wisconsinblock20}
\alias{wisconsinblock20}
\title{Wisconsin}
\format{
A data frame with 4692 rows and 398 variables:
}
\source{
/url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(georgiablock20)
}
\description{
Blocks for the state of Wisonsin
}
\keyword{datasets}
