% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wyomingblock20.R
\docType{data}
\name{wyomingblock20}
\alias{wyomingblock20}
\title{Wyoming}
\format{
A data frame with 457 rows and 398 variables:
}
\source{
/url{https://www2.census.gov/programs-surveys/decennial/2020/data/01-Redistricting_File--PL_94-171/}
}
\usage{
data(wyomingblock20)
}
\description{
Blocks for the state of Wyoming
}
\keyword{datasets}
